pyproject.toml
[tool.poetry]
name = "rigol-dashboard"
version = "0.1.0"
description = "Display waveforms from Rigol DS1054Z"
authors = ["DavidVentura <davidventura27@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
plotly-express = "^0.4.1"
pandas = "^2.1.4"
Pillow = "^10.1.0"
dash = "^2.14.2"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
mypy = "^1.8.0"
pytest = "^7.4.3"
pandas-stubs = "^2.1.4.231218"
types-pillow = "^10.1.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
